+++ examples/python/sumgui/demo_report_dashboard_seaborn.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+from demo_report_dashboard import main;
+
+if __name__ == "__main__":
+    main("seaborn");